Fix display job details at job details page
