+++ components/commons/user/item-jobdetails/style.module.scss
@@ -35,6 +35,9 @@
 .companyInfo div {
   margin-left: 20px;
 }
+.imgCompany {
+  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.05);
+}
 
 .jobTitle {
   margin-top: 20px;
